requirements.txt
web3==5.31.3
python-dotenv==1.0.0
py-solc-x==1.1.1